user_interface/requirements.txt
pyautogen[retrievechat]==0.1.14
openai==0.28.1
panel==1.3.6
chromadb==0.4.15